+++ game/exceptions.py
@@ -1,0 +1,6 @@
+class InactiveCompetitionError(Exception):
+    pass
+
+
+class CannotLeaveCompetitionError(Exception):
+    pass
