+++ game/models/__init__.py
@@ -1,0 +1,3 @@
+from .competition import Competition  # NOQA
+from .game import Game  # NOQA
+from .score import HistoricalScore, Score  # NOQA
