+++ game/stats.py
@@ -1,5 +1,7 @@
 from collections import defaultdict, OrderedDict
 from itertools import groupby
+import json
+import operator
 
 from django.core.exceptions import ObjectDoesNotExist
 
@@ -153,3 +155,103 @@ def get_stats_per_week(player, games):
         )
 
     return sorted(stats_per_week.items())
+
+
+def get_latest_results_by_player(competition, nb_games, offset=0,
+                                 return_json=False):
+    """
+    Get nb_games latest scores for each player
+
+    :param nb_games:int number of games
+    :param return_json:boolean
+    :return:dict Dict with key=player and value=list of score objects
+
+    {player_a: [{skill: xx, played: xx, game: game_id}, ...]}
+    """
+    # add start to nb_games because slicing want the end position
+    nb_games += offset
+    games = (competition.games
+             .order_by('-id')
+             .prefetch_related('historical_scores')[offset:nb_games])
+
+    players = competition.get_players()
+    scores_by_player = {}
+
+    for game in reversed(games):
+        all_skills_by_game = {}
+
+        for player in players:
+            player_scores = scores_by_player.get(player, [])
+            result = {'game': game.id}
+
+            if player.id in [game.winner_id, game.loser_id]:
+                player_historical_score = None
+                # We don't use get() here so we don't hit the database
+                # since historical scores are prefetched
+                for historical_score in game.historical_scores.all():
+                    if (historical_score.player_id == player.id and
+                            historical_score.competition_id == competition.id):
+                        player_historical_score = historical_score
+                        break
+
+                assert player_historical_score is not None
+
+                result['skill'] = player_historical_score.score
+
+                result['win'] = player.id == game.winner_id
+                result['played'] = True
+            else:
+                result['played'] = False
+                if len(player_scores) == 0:
+                    result['skill'] = competition.get_last_score_for_player(
+                        player, game
+                    )
+                else:
+                    result['skill'] = player_scores[-1]['skill']
+
+            all_skills_by_game[player] = result['skill']
+
+            player_scores.append(result)
+            scores_by_player[player] = player_scores
+
+        positions_for_game = sorted(
+            all_skills_by_game.items(),
+            key=operator.itemgetter(1),
+            reverse=True
+        )
+        for idx, position in enumerate(positions_for_game, start=1):
+            scores_by_player[position[0]][-1]['position'] = idx
+
+    if return_json:
+        json_result = {}
+        for player, results in scores_by_player.items():
+            json_result[player.get_full_name()] = results
+
+        return json.dumps(json_result)
+
+    return scores_by_player
+
+
+def get_last_score_for_player(player, competition, default_score,
+                              last_game=None):
+    """
+    Return the latest historical score before game for a player in the
+    competition.
+
+    :param player:Player
+    :param game:Game
+    :param competition:Competition
+    :return:HistoricalScore
+    """
+    last_score = (competition.historical_scores.filter(player=player)
+                                               .order_by('-id'))
+
+    if last_game:
+        last_score = last_score.filter(game_id__lt=last_game.id)
+
+    last_score = last_score.first()
+
+    if last_score is None:
+        last_score = default_score
+
+    return last_score
